+++ libbpf-cargo/src/test.rs
@@ -18,6 +18,7 @@ use memmap2::Mmap;
 use tempfile::tempdir;
 use tempfile::NamedTempFile;
 use tempfile::TempDir;
+use test_log::test;
 
 use crate::build::build;
 use crate::gen::GenBtf;
@@ -139,17 +140,17 @@ fn test_build_default() {
     let (_dir, proj_dir, cargo_toml) = setup_temp_project();
 
     // No bpf progs yet
-    build(true, Some(&cargo_toml), None, Vec::new()).unwrap_err();
+    build(Some(&cargo_toml), None, Vec::new()).unwrap_err();
 
     // Add prog dir
     create_dir(proj_dir.join("src/bpf")).expect("failed to create prog dir");
-    build(true, Some(&cargo_toml), None, Vec::new()).unwrap_err();
+    build(Some(&cargo_toml), None, Vec::new()).unwrap_err();
 
     // Add a prog
     let _prog_file =
         File::create(proj_dir.join("src/bpf/prog.bpf.c")).expect("failed to create prog file");
 
-    build(true, Some(&cargo_toml), None, Vec::new()).unwrap();
+    build(Some(&cargo_toml), None, Vec::new()).unwrap();
 
     // Validate generated object file
     validate_bpf_o(proj_dir.as_path().join("target/bpf/prog.bpf.o").as_path());
@@ -167,7 +168,7 @@ fn test_build_invalid_prog() {
         File::create(proj_dir.join("src/bpf/prog.bpf.c")).expect("failed to create prog file");
     writeln!(prog_file, "1").expect("write to prog file failed");
 
-    build(true, Some(&cargo_toml), None, Vec::new()).unwrap_err();
+    build(Some(&cargo_toml), None, Vec::new()).unwrap_err();
 }
 
 #[test]
@@ -186,14 +187,14 @@ fn test_build_custom() {
         .expect("write to Cargo.toml failed");
 
     // No bpf progs yet
-    build(true, Some(&cargo_toml), None, Vec::new()).unwrap_err();
+    build(Some(&cargo_toml), None, Vec::new()).unwrap_err();
 
     // Add a prog
     create_dir(proj_dir.join("src/other_bpf_dir")).expect("failed to create prog dir");
     let _prog_file = File::create(proj_dir.join("src/other_bpf_dir/prog.bpf.c"))
         .expect("failed to create prog file");
 
-    build(true, Some(&cargo_toml), None, Vec::new()).unwrap();
+    build(Some(&cargo_toml), None, Vec::new()).unwrap();
 
     // Validate generated object file
     validate_bpf_o(
@@ -223,13 +224,13 @@ fn test_unknown_metadata_section() {
 
     // Add prog dir
     create_dir(proj_dir.join("src/bpf")).expect("failed to create prog dir");
-    build(true, Some(&cargo_toml), None, Vec::new()).unwrap_err();
+    build(Some(&cargo_toml), None, Vec::new()).unwrap_err();
 
     // Add a prog
     let _prog_file =
         File::create(proj_dir.join("src/bpf/prog.bpf.c")).expect("failed to create prog file");
 
-    build(true, Some(&cargo_toml), None, Vec::new()).unwrap();
+    build(Some(&cargo_toml), None, Vec::new()).unwrap();
 
     // Validate generated object file
     validate_bpf_o(proj_dir.as_path().join("target/bpf/prog.bpf.o").as_path());
@@ -241,15 +242,15 @@ fn test_enforce_file_extension() {
 
     // Add prog dir
     create_dir(proj_dir.join("src/bpf")).expect("failed to create prog dir");
-    build(true, Some(&cargo_toml), None, Vec::new()).unwrap_err();
+    build(Some(&cargo_toml), None, Vec::new()).unwrap_err();
 
     let _prog_file = File::create(proj_dir.join("src/bpf/prog_BAD_EXTENSION.c"))
         .expect("failed to create prog file");
-    build(true, Some(&cargo_toml), None, Vec::new()).unwrap_err();
+    build(Some(&cargo_toml), None, Vec::new()).unwrap_err();
 
     let _prog_file_again = File::create(proj_dir.join("src/bpf/prog_GOOD_EXTENSION.bpf.c"))
         .expect("failed to create prog file");
-    build(true, Some(&cargo_toml), None, Vec::new()).unwrap();
+    build(Some(&cargo_toml), None, Vec::new()).unwrap();
 }
 
 #[test]
@@ -257,7 +258,7 @@ fn test_build_workspace() {
     let (_dir, _, workspace_cargo_toml, proj_one_dir, proj_two_dir) = setup_temp_workspace();
 
     // No bpf progs yet
-    build(true, Some(&workspace_cargo_toml), None, Vec::new()).unwrap_err();
+    build(Some(&workspace_cargo_toml), None, Vec::new()).unwrap_err();
 
     // Create bpf prog for project one
     create_dir(proj_one_dir.join("src/bpf")).expect("failed to create prog dir");
@@ -269,7 +270,7 @@ fn test_build_workspace() {
     let _prog_file_2 = File::create(proj_two_dir.join("src/bpf/prog2.bpf.c"))
         .expect("failed to create prog file 2");
 
-    build(true, Some(&workspace_cargo_toml), None, Vec::new()).unwrap();
+    build(Some(&workspace_cargo_toml), None, Vec::new()).unwrap();
 }
 
 #[test]
@@ -286,7 +287,7 @@ fn test_build_workspace_collision() {
     let _prog_file_2 = File::create(proj_two_dir.join("src/bpf/prog.bpf.c"))
         .expect("failed to create prog file 2");
 
-    build(true, Some(&workspace_cargo_toml), None, Vec::new()).unwrap_err();
+    build(Some(&workspace_cargo_toml), None, Vec::new()).unwrap_err();
 }
 
 #[test]
@@ -300,16 +301,7 @@ fn test_make_basic() {
     let _prog_file =
         File::create(proj_dir.join("src/bpf/prog.bpf.c")).expect("failed to create prog file");
 
-    make(
-        true,
-        Some(&cargo_toml),
-        None,
-        Vec::new(),
-        true,
-        Vec::new(),
-        None,
-    )
-    .unwrap();
+    make(Some(&cargo_toml), None, Vec::new(), true, Vec::new(), None).unwrap();
 
     // Validate generated object file
     validate_bpf_o(proj_dir.as_path().join("target/bpf/prog.bpf.o").as_path());
@@ -338,7 +330,6 @@ fn test_make_workspace() {
         .expect("failed to create prog file 2");
 
     make(
-        true,
         Some(&workspace_cargo_toml),
         None,
         Vec::new(),
@@ -394,16 +385,7 @@ fn build_rust_project_from_bpf_c_impl(bpf_c: &str, rust: &str, run: bool) {
     // Lay down the necessary header files
     add_vmlinux_header(&proj_dir);
 
-    make(
-        true,
-        Some(&cargo_toml),
-        None,
-        Vec::new(),
-        true,
-        Vec::new(),
-        None,
-    )
-    .unwrap();
+    make(Some(&cargo_toml), None, Vec::new(), true, Vec::new(), None).unwrap();
 
     let mut cargo = OpenOptions::new()
         .append(true)
@@ -1217,7 +1199,7 @@ fn build_btf_mmap(prog_text: &str) -> Mmap {
     add_vmlinux_header(&proj_dir);
 
     // Build the .bpf.o
-    build(true, Some(&cargo_toml), None, Vec::new()).expect("failed to compile");
+    build(Some(&cargo_toml), None, Vec::new()).expect("failed to compile");
 
     let obj = OpenOptions::new()
         .read(true)