+++ libbpf-cargo/CHANGELOG.md
@@ -3,6 +3,8 @@ Unreleased
 - Removed `SkeletonBuilder::skip_clang_version_check`
 - Removed `--skip-clang-version-checks` option of `libbpf build`
   sub-command
+- Replaced `--debug` option of `libbpf` sub-command with `-v` /
+  `--verbose`
 
 
 0.25.0-beta.1